moved logic to hook
