+++ frontend/src/Hooks/useGameState.js
@@ -1,0 +1,347 @@
+import { constructPlayersOnTrialStatus, nightTimeStatus } from '../GameUtils';
+import React, { useContext, useReducer, useEffect, useCallback } from 'react';
+import { GeneralContext } from '../App';
+import socket from '../Socket';
+
+const initialState = {
+    screen: 'entry',
+    dayPeriod: 'day', // night or day
+    dayNumber: 1, // what day it is
+    alivePlayers: [],
+    status: '',
+    winningRole: '',
+    winners: [],
+    phase: '',
+    votingState: {
+        type: '', // role or discussion or trial or undefined
+        votablePlayers: [], // what other players can we vote for
+        vote: '', // my current Vote
+        playersWhoVoted: [], // other players who have voted
+        killedPlayer: '', // the outcome of the vote
+        timeToVote: '', // time to vote
+    },
+};
+
+const reducer = (state, action) => {
+    switch (action.type) {
+        case 'init': {
+            return {
+                ...state,
+                alivePlayers: [...action.alivePlayers],
+                ...(action.checkedPlayers && {
+                    checkedPlayers: action.checkedPlayers,
+                }),
+            };
+        }
+
+        case 'night-start': {
+            return {
+                ...state,
+                phase: 'night-start',
+                status: action.status,
+                dayPeriod: 'night',
+                screen: 'core',
+                votingState: {
+                    ...state.votingState,
+                    type: 'role',
+                    votablePlayers: [...action.votablePlayers],
+                    timeToVote: action.timeToVote,
+                },
+            };
+        }
+
+        case 'night-end': {
+            return {
+                ...state,
+                phase: 'night-end',
+                status: action.status,
+                alivePlayers: state.alivePlayers.filter(
+                    (p) => p !== action.playerKilled
+                ),
+                votingState: {
+                    ...initialState.votingState,
+                },
+            };
+        }
+
+        case 'day-start': {
+            return {
+                ...state,
+                phase: 'day-start',
+                dayPeriod: 'day',
+                dayNumber: state.dayNumber + 1,
+                status: action.status,
+                votingState: {
+                    ...state.votingState,
+                    timeToVote: action.timeToVote,
+                    type: 'discussion',
+                    votablePlayers: [...action.votablePlayers],
+                },
+            };
+        }
+        // Look into this
+        case 'discussion-end': {
+            return {
+                ...state,
+                phase: 'discussion-end',
+                status: action.status,
+                votingState: {
+                    ...initialState.votingState,
+                    type: 'trial',
+                    ...(action.votablePlayers.length && {
+                        votablePlayers: action.votablePlayers,
+                    }),
+                },
+            };
+        }
+
+        case 'trial-start': {
+            return {
+                ...state,
+                phase: 'trial-start',
+                status: action.status,
+                votingState: {
+                    ...state.votingState,
+                    timeToVote: action.timeToVote,
+                },
+            };
+        }
+        case 'trial-end': {
+            return {
+                ...state,
+                phase: 'trial-start',
+                status: action.status,
+                alivePlayers: state.alivePlayers.filter(
+                    (p) => p !== action.playerKilled
+                ),
+                votingState: {
+                    ...initialState.votingState,
+                },
+            };
+        }
+
+        case 'game-over': {
+            return {
+                ...state,
+                phase: 'game-over',
+                status: `${action.winningRole} wins!`,
+                winningRole: action.winningRole,
+                winners: [...action.winners],
+            };
+        }
+
+        case 'vote-update': {
+            return {
+                ...state,
+                votingState: {
+                    ...state.votingState,
+                    playersWhoVoted: action.playersWhoVoted,
+                },
+            };
+        }
+
+        case 'vote-update': {
+            return {
+                ...state,
+                votingState: {
+                    ...state.votingState,
+                    playersWhoVoted: action.playersWhoVoted,
+                },
+            };
+        }
+
+        case 'suspect-reveal': {
+            return {
+                ...state,
+                checkedPlayers: [...state.checkedPlayers, action.checkedPlayer],
+            };
+        }
+
+        // case 'update-status':
+        //     return {
+        //         ...state,
+        //         status: action.status,
+        //     };
+        // case 'kill-player':
+        //     return {
+        //         ...state,
+        //         alivePlayers: state.alivePlayers.filter(
+        //             (p) => p !== action.playerKilled
+        //         ),
+        //     };
+        // case 'increment-time':
+        //     return {
+        //         ...state,
+        //         dayPeriod: state.dayPeriod === 'day' ? 'night' : 'day',
+        //         dayNumber:
+        //             state.dayPeriod === 'day'
+        //                 ? state.dayNumber
+        //                 : state.dayNumber + 1,
+        //     };
+        // case 'winner':
+        //     return {
+        //         ...state,
+        //         winningRole: action.winningRole,
+        //         winners: [...action.winners],
+        //         screen: 'end',
+        //     };
+        // case 'check-player': {
+        //     return {
+        //         ...state,
+        //         checkedPlayers: [...state.checkedPlayers, action.checkedPlayer],
+        //     };
+        // }
+    }
+};
+
+export default function useGameState() {
+    const [state, dispatch] = useReducer(reducer, initialState);
+    const { state: generalState } = useContext(GeneralContext);
+
+    /**
+     * This effect will run on first render to initialise the alive players.
+     * We may need to lie about the dependency so that players dont just 'leave'
+     */
+    useEffect(() => {
+        const { role } = generalState;
+        const extraRoleState =
+            role === 'detective' ? { checkedPlayers: [] } : {};
+
+        dispatch({
+            type: 'init',
+            alivePlayers: generalState.players,
+            ...extraRoleState,
+        });
+    }, [generalState]);
+
+    useEffect(() => {
+        function onNightStart({ timeToVote }) {
+            const { role } = generalState;
+
+            const status = nightTimeStatus[role];
+
+            const votablePlayers = {
+                mafia: state.alivePlayers.filter(
+                    (p) => p !== generalState.nickname
+                ),
+                // TODO fix detective
+                detective: state.alivePlayers.filter(
+                    (p) =>
+                        p !== generalState.nickname &&
+                        state.checkedPlayers &&
+                        !state.checkedPlayers.some((c) => c.nickname === p)
+                ),
+                medic: state.alivePlayers,
+                civilian: [],
+            }[role];
+
+            dispatch({
+                type: 'night-start',
+                status,
+                votablePlayers,
+                timeToVote,
+            });
+
+            // dispatch({ type: 'change-screen', screen: 'core' });
+        }
+
+        function onNightEnd({ playerKilled }) {
+            dispatch({
+                type: 'night-end',
+                status: `${playerKilled} was killed in the night`,
+                playerKilled,
+            });
+        }
+
+        function onDayStart({ timeToVote }) {
+            dispatch({
+                type: 'day-start',
+                status: 'Please vote for someone',
+                votablePlayers: state.alivePlayers.filter(
+                    (p) => p !== generalState.nickname
+                ),
+                timeToVote,
+            });
+        }
+
+        function onDiscussionEnd({ playersOnTrial }) {
+            dispatch({
+                type: 'discussion-end',
+                status: constructPlayersOnTrialStatus(playersOnTrial),
+                votablePlayers: !playersOnTrial.includes(generalState.nickname)
+                    ? playersOnTrial
+                    : [],
+            });
+        }
+
+        function onTrialStart({ timeToVote }) {
+            dispatch({
+                type: 'trial-start',
+                status: state.votingState.votablePlayers.length
+                    ? 'Please vote for someone'
+                    : 'You are on trial',
+                timeToVote,
+            });
+        }
+
+        function onTrialEnd({ playerKilled }) {
+            dispatch({
+                type: 'trial-end',
+                status: `${playerKilled} was hanged`, // TODO if playerKilled === null
+                playerKilled,
+            });
+        }
+
+        function onGameOver({ winningRole, winners }) {
+            dispatch({
+                type: 'game-over',
+                winningRole: winningRole.toLowerCase(),
+                winners,
+            });
+        }
+
+        function onVoteUpdate({ voteMap }) {
+            dispatch({
+                type: 'vote-update',
+                playersWhoVoted: Object.keys(voteMap),
+            });
+        }
+
+        function onSuspectReveal(checkedPlayer) {
+            dispatch({
+                type: 'suspect-reveal',
+                checkedPlayer,
+            });
+        }
+
+        socket.on('night-start', onNightStart);
+        socket.on('night-end', onNightEnd);
+        socket.on('day-start', onDayStart);
+        socket.on('discussion-end', onDiscussionEnd);
+        socket.on('trial-start', onTrialStart);
+        socket.on('trial-end', onTrialEnd);
+        socket.on('game-over', onGameOver);
+
+        socket.on('day-vote-update', onVoteUpdate);
+        socket.on('trial-vote-update', onVoteUpdate);
+
+        socket.on('suspect-reveal', onSuspectReveal);
+
+        return () => {
+            socket.removeListener('night-start', onNightStart);
+            socket.removeListener('night-end', onNightEnd);
+            socket.removeListener('day-start', onDayStart);
+            socket.removeListener('discussion-end', onDiscussionEnd);
+            socket.removeListener('trial-start', onTrialStart);
+            socket.removeListener('trial-end', onTrialEnd);
+            socket.removeListener('game-over', onGameOver);
+
+            socket.removeListener('day-vote-update', onVoteUpdate);
+            socket.removeListener('trial-vote-update', onVoteUpdate);
+
+            socket.removeListener('suspect-reveal', onSuspectReveal);
+        };
+    }, [state]);
+
+    return state;
+}
